platform-api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
jinja2==3.1.2
pyyaml==6.0.1
requests==2.31.0
apscheduler==3.10.4
beautifulsoup4==4.12.2
docker==7.0.0
python-dateutil==2.8.2
